+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased] - 2016-06-06
+## [Unreleased]
 ### Added
 - Changelog
+- Custom types
 
 ## [0.2.13] - 2016-06-06
 ### Added